# doc/api_docs/openapi.bundled.yaml
openapi: 3.1.0
info:
  title: Treasure Doc API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Treasure Doc（个人博客 + 后台管理）的 HTTP API 规格。

    ## 文件组织
    本仓库的 API 规格采用「**一个端点一个文件**」的组织方式：

    - `openapi.yaml`（本文件）：聚合根。只声明 `info` / `servers` / `tags` / `security` 与全部 `paths`，
      每个 operation（`get` / `post` / `put` / `patch` / `delete`）通过 $ref 指向对应端点文件。
    - `blog/`（22 个）：公开只读博客 API `/api/blog/*` 与站点根 `robots.txt` / `sitemap.xml` / `rss.xml`。
    - `blog-mgr/`（80 个）：后台管理 API `/api/blog-mgr/*`（含 `/resources/<资源>/` 子目录、备份、媒体库、编辑历史）。
    - `memo/`（7 个）：前台速记本 API `/api/memo/*`（登录即博主本人）。
    - `publish/`（12 个）：机器令牌发布 API `/api/publish/*`。
    - `user/`（5 个）：鉴权 API `/api/user/*` 与健康检查 `/ping`。
    - `backup/`（1 个）：NAS 机器令牌导出 `/api/backup/export`。
    - `components/`：跨端点共享的 `schemas`、`responses`、`parameters`，端点文件用相对路径引用。

    端点文件的内容是一个 **operation 对象**（不含 `openapi:` / `paths:` 包装），首行注释写明
    「HTTP 方法 + 完整路径 + 所属模块 + 路由注册位置」。

    ## 解析方式
    这是「subdocuments」风格的多文件规格，**不能**被 Swagger UI 直接加载，需先打包：

    ```bash
    npx @redocly/cli bundle doc/api_docs/openapi.yaml -o /tmp/openapi.bundled.yaml
    ```

    然后即可用 Swagger UI / Redoc / `npx @redocly/cli preview-docs doc/api_docs/openapi.yaml` 浏览。

    ## 统一响应约定
    - 除 `GET /ping` 与站点根三个文件外，所有接口都返回 `{code, msg, data}` 信封。
    - 业务成功与失败通常都返回 HTTP 200，靠 `code` 区分（`0` = 成功）；
      认证/权限/限流中间件则返回真实的 401 / 403 / 429。
...

    ## 鉴权
    | 请求头 | 适用接口 | 说明 |
    | --- | --- | --- |
...

...
servers:
...
security:
  - {}
...
paths:
...
  /api/publish/uploads:
    post:
      operationId: publishUploadMedias
      tags:
        - publish
      summary: 发布时上传图片 / 附件
      description: |
        用**发布令牌**鉴权（请求头 `X-Publish-Token`，不是管理员 `X-Token`），供发布插件在替换文章内图片/附件时上传文件。

        单次最多 **30 个**文件、单文件**上限 100MB**。图片/视频按**内容嗅探**判定（防伪装），PDF 也按内容识别；
        其余附件（docx / xlsx / zip / pptx / ppt / txt / md / csv 等）按原始后缀白名单保留语义。

        仍按内容 sha256 命名去重，返回可直接替换引用的 `/files/blog/<sha256>.<ext>` 地址；相同内容重复上传会复用已有文件
        （`existed: true`，服务端不重复写盘）。返回的 `list` **顺序与提交顺序一致**。

        **任一文件失败则整体失败**，HTTP 仍为 200 但响应体 `code` 为 `1`、`msg` 会带上出错文件名（见失败示例）。
      security:
        - XPublishToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  maxItems: 30
                  items:
                    type: string
                    format: binary
                  description: 待上传的图片 / 附件（单次最多 30 个，单个上限 100MB）
                file:
                  type: string
                  format: binary
                  description: 兼容单数字段；`files` 为空时回退使用该字段（可选）
            encoding:
              files:
                contentType: image/png, image/jpeg, image/gif, image/webp, image/bmp, video/mp4, video/webm, video/quicktime, video/x-msvideo, video/x-matroska, video/mpeg, application/pdf
      responses:
        '200':
          description: |
            成功时 code 为 0、data 为 `{list:[...]}`（顺序与提交顺序一致）；失败时 code 为 1、msg 含文件名与原因、data 为 {}，
            HTTP 仍为 200。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/ManageUploadMediaList'
                          - $ref: '#/components/schemas/EmptyData'
              examples:
                success:
                  value:
                    code: 0
                    msg: 操作成功
                    data:
                      list:
                        - path: /files/blog/9f2c8b1d6e4a0f3c5d7b9a1e2f4c6d8b0a2c4e6f8a1b3d5c7e9f0a2b4c6d8e1f.png
                          name: 9f2c8b1d6e4a0f3c5d7b9a1e2f4c6d8b0a2c4e6f8a1b3d5c7e9f0a2b4c6d8e1f.png
                          size: 182734
                          existed: false
                        - path: /files/blog/1a2b3c4d5e6f708192a3b4c5d6e7f8091a2b3c4d5e6f708192a3b4c5d6e7f809.pdf
                          name: 1a2b3c4d5e6f708192a3b4c5d6e7f8091a2b3c4d5e6f708192a3b4c5d6e7f809.pdf
                          size: 5242880
                          existed: true
                failure:
                  value:
                    code: 1
                    msg: '[a.png]文件大小不能超过 100MB'
                    data: {}
        '401':
          $ref: '#/components/responses/PublishUnauthorized'
        '403':
          $ref: '#/components/responses/PublishIPForbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/user/captcha:
    get:
      operationId: getUserCaptcha
      tags:
        - user
      summary: 获取图形验证码
      description: |
        返回一张**算术题**验证码。`captcha` 是 base64 图片数据（可直接作为 `<img src>`），
        `captchaId` 需在调用 `POST /api/user/login` 时回传。

        验证码为**一次性消费**：校验通过后立即失效。

        生成失败时 code 为 1、msg「验证码生成失败」、data 为 {}（该模块沿用「失败也返回 HTTP 200」的早期约定）。
      security:
        - {}
      responses:
        '200':
          description: |
            成功时 code 为 0、data 为验证码；失败时 code 为 1、msg 为具体原因（如「验证码生成失败」）、data 为 {}。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/UserCaptcha'
                          - $ref: '#/components/schemas/EmptyData'
              examples:
                success:
                  value:
                    code: 0
                    msg: 操作成功
                    data:
                      captchaId: 9Xk2pQ7fLm3Tz1aB
                      captcha: data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAIAAAACACAYAAADDPmHLAAAA...
                failure:
                  value:
                    code: 1
                    msg: 验证码生成失败
                    data: {}
  /api/user/login:
    post:
      operationId: userLogin
      tags:
        - user
      summary: 用户登录
      description: |
        `account` 支持账号或邮箱（大小写不敏感）。验证码是否必填取决于服务端配置 `captcha.enabled`，
        开启时缺失或错误会返回验证码错误码 40101。

        **同一账号最多保留 3 个登录令牌**，超出时最早的一个会被删除。

        成功后返回的 `token` 用于后续请求的 `X-Token` 头。`requirePwdReset` 为 true 时前端提醒一次修改密码，
        鉴权中间件**不再拦截**管理接口。

        **限流**：该路径受令牌桶限流（0.1 次/秒，容量 5，按客户端 IP），超限返回 HTTP 429（见 429 响应）。
      security:
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginInput'
      responses:
        '200':
          description: |
            成功时 code 为 0、data 为登录用户信息；失败时 code 为 1（或验证码错误 40101）、
            msg 为具体原因（如「密码或账号(邮箱)不能为空」「账号 xxx 没有找到」「账号不可用或未激活」「获取用户token失败」）、
            data 为 {}（该模块沿用「失败也返回 HTTP 200」的早期约定）。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/UserLoginResult'
                          - $ref: '#/components/schemas/EmptyData'
              examples:
                success:
                  value:
                    code: 0
                    msg: 操作成功
                    data:
                      id: '1930526412345678901'
                      nickname: 管理员
                      account: admin
                      email: admin@example.com
                      userType: 100
                      userStatus: 1
                      token: 8f2c8b1d6e4a0f3c5d7b9a1e2f4c6d8b0a2c4e6f8a1b3d5c7e9f0a2b4c6d8e1f
                      requirePwdReset: false
                failure:
                  value:
                    code: 1
                    msg: 账号或密码错误
                    data: {}
                captchaFailure:
                  value:
                    code: 40101
                    msg: 验证码错误或已失效
                    data: {}
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/user/logout:
    post:
      operationId: userLogout
      tags:
        - user
      summary: 退出登录
      description: |
        删除当前 `X-Token` 对应的登录令牌。

        令牌无效或已过期时由鉴权中间件返回 HTTP 401；成功时服务端返回空对象，data 为 {}。
      security:
        - XToken: []
      responses:
        '200':
          description: |
            成功时 code 为 0、data 为 {}；失败时 code 为 1、msg 为具体原因、data 为 {}（该模块沿用「失败也返回 HTTP 200」的早期约定）。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EmptyData'
              examples:
                success:
                  value:
                    code: 0
                    msg: 操作成功
                    data: {}
                failure:
                  value:
                    code: 1
                    msg: 退出登录失败
                    data: {}
        '401':
          $ref: '#/components/responses/UserUnauthorized'
  /api/user/change-pwd:
    post:
      operationId: userChangePassword
      tags:
        - user
      summary: 修改当前用户密码
      description: |
        修改当前登录用户的密码，需要校验原密码。

        新密码长度 **8–16 位**，`rePassword` 必须与 `password` 一致。修改成功后服务端会清除该用户的
        `requirePwdReset` 标记。

        请求体校验失败（如字段缺失、密码长度不符）时由 binding 直接拒绝，返回 code 1 与具体 msg。
      security:
        - XToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserChangePwdInput'
      responses:
        '200':
          description: |
            成功时 code 为 0、data 为 {}；失败时 code 为 1、msg 为具体原因（如「原密码错误」）、data 为 {}
            （该模块沿用「失败也返回 HTTP 200」的早期约定）。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EmptyData'
              examples:
                success:
                  value:
                    code: 0
                    msg: 操作成功
                    data: {}
                failure:
                  value:
                    code: 1
                    msg: 原密码错误
                    data: {}
        '401':
          $ref: '#/components/responses/UserUnauthorized'
...
  /api/backup/export:
    get:
      operationId: nasExportBackup
      tags:
        - backup
      summary: NAS 机器令牌导出完整备份
      description: |
        **免登录**的机器令牌接口，用 `X-Backup-Token` 鉴权（令牌取值来源：后台「备份令牌」配置，
        未配置时回退配置文件 `[backup].apiToken`；两者都为空则一律 401）。

        **只读**：不提供导入/删除能力，适合定时任务拉取。

        可叠加 IP 白名单（配置项 `backup.allowIPs`，支持精确 IP 与 CIDR，未配置时不做限制）。

        仅 SQLite 驱动支持备份，其余驱动返回失败（msg「仅 SQLite 驱动支持备份」）。
      security:
        - XBackupToken: []
      responses:
        '200':
          description: |
            成功时现场打包出一个 `.tar.gz` 完整备份包（数据库 + `files/` 上传文件 + manifest）并流式返回，
            响应头带 `Content-Disposition: attachment; filename="<name>"`，Content-Type 为 `application/gzip` 或 `application/octet-stream`。
            失败时返回 JSON 信封 `{code: 1, msg: "..." , data: {}}`（如「仅 SQLite 驱动支持备份」「创建临时目录失败」）。
          headers:
            Content-Disposition:
              schema:
                type: string
              description: 附件文件名，形如 `attachment; filename="treasure-backup-20260301.tar.gz"`
          content:
            application/gzip:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '401':
          $ref: '#/components/responses/BackupUnauthorized'
        '403':
          $ref: '#/components/responses/BackupIPForbidden'
components:
  securitySchemes:
    XToken:
      type: apiKey
      in: header
      name: X-Token
      description: 登录成功后签发的用户令牌；后台接口还要求 `userType` ∈ {2, 100}。
    XPublishToken:
      type: apiKey
      in: header
      name: X-Publish-Token
      description: 机器令牌；取值为后台「站点设置 → 发布令牌」配置的值，未配置时所有 `/api/publish/*` 请求返回 401。
    XBackupToken:
      type: apiKey
      in: header
      name: X-Backup-Token
      description: NAS 机器令牌；取值为后台配置的备份令牌（回退配置文件 `[backup].apiToken`）。
  schemas:
    Envelope:
      type: object
      description: |
        统一响应信封，所有接口（除 GET /ping 与 robots/sitemap/rss 外）都返回该结构。
        业务结果由 code 表达：0 = 成功，其余为各模块业务码。失败时 data 通常为 null，
        字段级校验失败时 data 为 `{"field": "..."}`（管理端）或 `{}`（用户模块）。
      required:
        - code
        - msg
        - data
      properties:
        code:
          type: integer
          description: 业务码，0 表示成功
          example: 0
        msg:
          type: string
          description: 提示信息；成功时为空字符串
          example: ''
        data:
          description: |
            业务数据；失败时通常为 null。各端点在此声明具体结构
            （对象、数组或 null）；未声明的端点表示任意值。
            此处不限定类型，以便端点用 allOf 追加自己的 data 结构。
...
    EmptyData:
      type: object
      maxProperties: 0
      description: |
        失败响应中的空对象 `data`。用户模块与上传/备份类端点沿用早期约定：
        失败时 HTTP 仍为 200，`code: 1`、`msg` 为具体原因、`data` 固定为 `{}`。
        这些端点的 `data` 声明为「成功结构 或 EmptyData」的 oneOf。
    ManageUploadedMedia:
      type: object
      description: 单个媒体文件上传结果
      required:
        - path
        - name
        - size
        - existed
      properties:
        path:
          type: string
          example: /files/blog/9f2c8b1d....png
        name:
          type: string
          description: 存储文件名
        size:
          type: integer
          format: int64
        existed:
          type: boolean
          description: true 表示服务端已存在相同内容（sha256 相同），未重复写盘
    ManageUploadMediaList:
      type: object
      description: 多文件上传结果，顺序与提交顺序一致
      required:
        - list
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/ManageUploadedMedia'
...
    UserCaptcha:
      type: object
      description: 图形验证码（算术题）
      required:
        - captchaId
        - captcha
      properties:
        captchaId:
          type: string
          description: 验证码 ID，登录时以 `captchaId` 回传
        captcha:
          type: string
          description: base64 图片数据（可直接作为 `<img src>`）
    UserLoginInput:
      type: object
      description: |
        登录请求。`account` 字段同时支持账号与邮箱（均唯一）。
        验证码是否必填取决于服务端 `captcha.enabled` 配置，因此 `captchaId` / `verifyCode`
        在结构上没有 required 约束，由服务层按配置校验。
      required:
        - account
        - password
      properties:
        account:
          type: string
          minLength: 6
          maxLength: 100
          description: 账号或邮箱
        password:
          type: string
          minLength: 6
          maxLength: 100
        captchaId:
          type: string
          maxLength: 100
          description: 验证码 ID，取自 `GET /api/user/captcha` 的 `captchaId`
        verifyCode:
          type: string
          maxLength: 100
          description: 用户输入的验证码答案（算术题结果）
    UserLoginResult:
      type: object
      description: 登录成功返回的当前用户
      required:
        - id
        - nickname
        - account
        - email
        - userType
        - userStatus
        - token
      properties:
        id:
          type: string
        nickname:
          type: string
        account:
          type: string
        email:
          type: string
        password:
          type: string
          description: 密码哈希（历史遗留字段，随模型一并返回）
        userType:
          type: integer
          description: 1 普通用户、2 管理员、100 超级管理员
          enum:
            - 1
            - 2
            - 100
        userStatus:
          type: integer
          description: 1 可用、2 不可用、3 未激活
        mobile:
          type: string
        avatar:
          type: string
        bio:
          type: string
        currentRoomId:
          type: string
        requirePwdReset:
          type: boolean
          description: 为 true 时前端登录后提醒一次修改密码（不拦截接口）
        token:
          type: string
          description: 后续请求的 X-Token
        createdAt:
          type: string
          format: date-time
          description: 注意：用户模块 BaseModel 的字段名与列名互换，createdAt 实际取自 updated_at 列
        updatedAt:
          type: string
          format: date-time
          description: 注意：实际取自 created_at 列
        deletedAt:
          description: 软删除时间；未删除时为 null
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
    UserChangePwdInput:
      type: object
      description: 修改当前登录用户的密码，需要校验原密码
      required:
        - oldPassword
        - password
        - rePassword
      properties:
        oldPassword:
          type: string
          description: 原密码
        password:
          type: string
          minLength: 8
          maxLength: 16
          description: 新密码，8–16 位
        rePassword:
          type: string
          minLength: 8
          maxLength: 16
          description: 重复新密码，必须与 password 一致
...
  responses:
...
    RateLimited:
      description: |
        触发限流（HTTP 429）。仅作用于 `/api/user/login`（0.1/s，容量 5）、
        `/api/blog-mgr/uploads/`（0.5/s，容量 20）、`/api/publish`（0.2/s，容量 20），按客户端 IP 令牌桶计数。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          example:
            code: 42900
            msg: 请求过于频繁，请稍后再试
            data: null
    PublishUnauthorized:
      description: 发布令牌未配置或无效（HTTP 401，code 40100）；请求头 `X-Publish-Token` 必须等于后台配置的发布令牌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          examples:
            missing:
              value:
                code: 40100
                msg: 未配置发布令牌
                data: null
            invalid:
              value:
                code: 40100
                msg: 发布令牌无效
                data: null
    PublishIPForbidden:
      description: 来源 IP 不在白名单内（HTTP 403，code 40300）；仅当配置了 `publish.allowIPs` 时校验
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          example:
            code: 40300
            msg: 来源 IP 不在白名单内
            data: null
    UserUnauthorized:
      description: 未登录或登录已失效（HTTP 401，code 1）；`X-Token` 缺失或已过期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          examples:
            missing:
              value:
                code: 1
                msg: 请先登录
                data: null
            invalid:
              value:
                code: 1
                msg: 认证失败，请重新登录
                data: null
    BackupUnauthorized:
      description: NAS 备份令牌未配置或无效（HTTP 401，code 40100）；请求头 `X-Backup-Token` 必须等于后台配置的备份令牌（回退配置文件 `[backup].apiToken`）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          examples:
            missing:
              value:
                code: 40100
                msg: 未配置 NAS 备份令牌
                data: null
            invalid:
              value:
                code: 40100
                msg: NAS 备份令牌无效
                data: null
    BackupIPForbidden:
      description: 来源 IP 不在白名单内（HTTP 403，code 40300）；仅当配置了 `backup.allowIPs` 时校验
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          example:
            code: 40300
            msg: 来源 IP 不在白名单内
            data: null
...